Fucked Up Evil Update
was following a tutorial to get a better options menu set up, and something the tutorial said would work just doesnt and sunk cost fallacy is keeping me from just undo-ing it all so heres the updated non-functioning code for the me in probably a few weeks to fix.

will be posting a build with working multiplayer tho, since that works fine. just disabling the options menu for now.