Options Implemented (sorta)
i got it bandaid fixed but theres a lot of redundant stuff i think. also, a few bugs and controls dont save once you leave the game. i was going to put that in but it was looking like a lot and i wanted to make sure i put this up before i broke some shit, so you might see leftovers from that if you go poking around.